Do not only trust GitHub maintainer_can_modify field
This implements a new field called `base_is_modifiable` which shows if the base
of a PR can be modified for the strict workflow. The current field provided by
GitHub does not cover the case where the base branch comes from the same repo
than the target branch, and not from a fork — see #51. This should fix it.

Fixes #51